Created Abstract Player Class, transfered User and Dealer code there, getDealerHand and getUserHand Renamed... Refracting all tetss listed that would have have those names.
