+++ src/main/java/Text_Based_UI/Display.java
@@ -10,7 +10,7 @@ public class Display {
         System.out.println(GameMessages.outline);
         System.out.println(GameMessages.dealerPoints(dealCard.getDealer().getPoints()));
         System.out.println(GameMessages.examCard);
-        System.out.println(GameMessages.dealerCards(dealCard.getDealer().getDealerHand().size()));
+        System.out.println(GameMessages.dealerCards(dealCard.getDealer().getHand().size()));
 
 
         System.out.println(GameMessages.userPoints(dealCard.getUser().getPoints()));
